tests(front): reflect job components refactoring
